refactor(sdk): move generator sources under src/
