+++ apps/sdk-generator/src/ownership.ts
