+++ apps/sdk-generator/src/project.ts
